app/requirements.txt
requests==2.18.4
psycopg2-binary
pandas
unittest
connexion==2.2.0
Flask==0.12.2
Werkzeug==0.14.1
SQLAlchemy==1.1.11
geopandas==0.5.0
PyYAML==5.1.1
